requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
duckdb>=0.10.0
pandas>=2.0.0
pydantic>=2.0.0
httpx>=0.25.0
pyyaml>=6.0
openai>=1.0.0
